+++ src/components/Backward.tsx
@@ -1,10 +1,9 @@
 import React, { useContext } from 'react'
 import { DEFAULT_POPUP_PAGE_TYPE } from '../consts'
-import { DisplayPageContext } from '../popup/Popup'
+import { DisplayPageContext } from './DisplayPageContextProvider'
 
 const Backward = () => {
   const { setDisplayPageType } = useContext(DisplayPageContext)
-
   return (
     <button
       className="absolute block top-1 left-1 w-4 h-4 icon-color border-b-2 border-l-2 rotate-45"
